Add files via upload
This world is filled with an immense amount of data and that is increasing every day. Go
through the all data is a very time-Consuming Process. There is a great need to reduce
much of this text data to shorter, focused summaries that capture the salient details, so we
can navigate it more effectively as well as check whether the larger documents contain the
information that we are looking for.
To solve this problem many Audio summarization tools are available in the market but some
of the features are yet to be introduced in them. We are effective, easier, and more timesaving for the users.
Proposed system is much more helpful to easily understand the large text by summary. The
input data is processed using NLP and the processed data input is converted into the vector
using word embedding. Word embedding means a set of feature learning and language
modeling techniques in NLP where phrases or words from the vocabulary are mapped to
vectors. Sentence ranking is used to extract higher-ranked sentences from sentences and
forms the extractive summary of the news text.
The user will be able to upload the Content in Text and Audio format for Summarization.
The user will be able to save the summarized file. The user will be able to visualize given
Data.